server/requirements.txt
pytest>=8.3.4
Flask-Testing>=0.8.1
flask~=3.1.0
flask-jwt-extended~=4.7.1
pyyaml>=6.0.2
app~=0.0.1
Werkzeug~=3.1.3
pymongo>=4.11.2
spotipy>=2.25.1